update utility functions

+++ include/str_utils.h
@@ -13,7 +13,7 @@ typedef enum{
 char *make_str(char *str, char c, size_t len);
 
 //--- Getters ---
-unsigned int get_size(char *str);
+size_t get_size(char *str);
 
 //--- Conversion ---
 char* itoa(unsigned int val, char *str, radix rad);
